#62 - 찜 기능

+++ src/main/java/com/example/travel/repository/PlaceRepository.java
@@ -4,8 +4,10 @@ import com.example.travel.entity.Place;
 import org.springframework.data.domain.Page;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.domain.Pageable;
-public interface PlaceRepository extends JpaRepository<Place, Integer> {
+
+public interface PlaceRepository extends JpaRepository<Place, Long> {
     Place findByContentId(int contentId);
     Page<Place> findByTitleContaining(String title, Pageable pageable);
+
 }
